fix(governance): close APP-39 by migrating money fields to Decimal [AITBC-88]
- Convert token_balance, staked_tokens, voting_power, rewards, stake amounts,
  treasury balances, vote power, and proposal vote tallies from float to
  Decimal(Numeric(20, 8)) to prevent precision drift on the money path.
- BlockchainClient.get_voting_power now returns Decimal.
- Service methods calculate voting power and manage stakes/delegations in
  Decimal; staking/delegation endpoint amounts are Decimal.
- Fix a shadowed variable in the /v1/transactions execute path.

+++ apps/governance/src/governance_service/clients/blockchain.py
@@ -14,6 +14,7 @@ from __future__ import annotations
 
 import json
 import logging
+from decimal import Decimal
 from typing import Any
 
 from aitbc.blockchain.rpc_client import BlockchainClient as BaseBlockchainClient
@@ -41,7 +42,7 @@ class BlockchainClient(BaseBlockchainClient):
     signing (secp256k1) and voting power queries.
     """
 
-    async def get_voting_power(self, address: str, chain_id: str | None = None) -> float:
+    async def get_voting_power(self, address: str, chain_id: str | None = None) -> Decimal:
         """Get on-chain voting power (balance) for an address.
 
         This is the on-chain balance snapshot used for vote weight.
